requirements.txt
discord.py==2.3.2
prometheus_client==0.20.0
PyNaCl==1.5.0
